// src/app/interfaces/IUsuario.ts
export interface Usuario {
    IDPER: number;
    NOMPER: String;
    APEPER: String;
    DNIPER: number;
/*  ESTPER: String;
    TIPPER: String; */
    DIRPER: String;
    GRAAUL: String;
    SECAUL: String;
    ASISTENCIAS: String;
    FALTAS: String;
    CB: String;
    MV: String;
    USUPER: String;
    CONTRPER?: String;
}